+++ src/main/java/org/uepb/algorithms/InsertionSort.java
@@ -2,11 +2,11 @@ package org.uepb.algorithms;
 
 public class InsertionSort extends SortAlgorithm {
     @Override
-    public int[] sort(int[] array) {
+    public double[] sort(double[] array) {
         int n = array.length;
 
         for (int j = 1; j < n; j++) {
-            int key = array[j];
+            double key = array[j];
             int i = j-1;
             while ( (i > -1) && ( array [i] > key ) ) {
                 array [i+1] = array [i];
